+++ src/modules/HomeModules/EventCardComponent/EventCardComponent.tsx
@@ -1,4 +1,4 @@
-import "../../../scss/upComingComponent.scss";
+import "../../../scss/UpComingComponent.scss";
 function EventCard() {
   return (
     <>
