+++ src/pages/HomePage/HomePage.tsx
@@ -1,5 +1,5 @@
 import HomeHero from "../../modules/HomeModules/HomeHero";
-import Upcoming from "../../modules/HomeModules/UpComingEvents/UpComingEvents";
+import Upcoming from "../../modules/HomeModules/UpComingEvents/UpComingEventsComponent";
 import Partners from "../../modules/HomeModules/PartnersComponent/Partner";
 
 function HomePage() {
